+++ src/test/java/domain/ParticipantTest.java
@@ -2,7 +2,6 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import constant.Rank;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
